+++ canvas/getShadowBlur.go
@@ -5,6 +5,6 @@ package canvas
 //
 // pt_br: Retorna o valor de borrão da sombra
 //     Valor padrão: 0
-func (el *Canvas) GetShadowBlur() float64 {
-	return el.SelfContext.Get("shadowBlur").Float()
+func (el *Canvas) GetShadowBlur() int {
+	return el.SelfContext.Get("shadowBlur").Int()
 }
