+++ canvas/setHeight.go
@@ -1,6 +1,6 @@
 package canvas
 
-func (el *Canvas) SetHeight(height float64) {
+func (el *Canvas) SetHeight(height int) {
 	el.SelfElement.Set("height", height)
 	el.SelfContext.Set("height", height)
 }
