+++ canvas/setWidth.go
@@ -1,6 +1,6 @@
 package canvas
 
-func (el *Canvas) SetWidth(width float64) {
+func (el *Canvas) SetWidth(width int) {
 	el.SelfElement.Set("width", width)
 	el.SelfContext.Set("width", width)
 }
